Changed the operating system used by the build job to `ubuntu-latest` instead of `ubuntu-18.04`.
Modified the code to install useful utilities using `apk` package manager in one workflow and `apt-get` package manager in another workflow.

Removed unnecessary whitespace in one of the workflows.

Changed the command used to check if a file exists from `if [ -f "$FILE" ]` to `if [ -f "$FILE" ]; then`.

Added a check to see if the file is copyrighted to the Pittsburgh Supercomputing Center in the README.md file.

Added checks to see if there exists at least one Singularity definition file, build script, remote build script, and test script.